pyproject.toml
[project]
name = "deal-deck-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "ruff>=0.14.7",
    "pytest>=8.0.0",
    "numpy>=1.26.0",
]
